requirements.txt
flask
opencv-python
datetime
pillow
numpy
pathlib
pandas
scikit-learn